requirements.txt
nats-py>=2.2.0
elasticsearch>=8.0.0
click>=8.1.0
rich>=13.0.0
pyyaml>=6.0
kubernetes>=26.1.0
python-dateutil>=2.8.2
jinja2>=3.1.0
pydantic>=2.0.0
setuptools>=65.0.0 